+++ library/Episciences/Review.php
@@ -63,6 +63,8 @@ class Episciences_Review
     public const SETTING_ALLOW_EDIT_VOLUME_TITLE_WITH_PUBLISHED_ARTICLES = 'allowEditVolumeTitleWithPublishedArticles';
     public const SETTING_ENCAPSULATE_REVIEWERS = 'encapsulateReviewers';
     public const SETTING_EDITORS_CAN_REASSIGN_ARTICLES = 'editorsCanReassignArticle';
+    public const SETTING_AUTHORS_CAN_CONTACT_EDITORS = 'authorsCanContactEditors';
+    public const SETTING_DISCLOSE_EDITOR_NAMES_TO_AUTHORS = 'discloseEditorNamesToAuthors';
     //Assignation automatique de rédacteurs
     public const SETTING_SYSTEM_AUTO_EDITORS_ASSIGNMENT = 'systemAutoEditorsAssignment';
     //Paramétrage avancé
@@ -208,6 +210,8 @@ class Episciences_Review
             self::SETTING_SPECIAL_ISSUE_ACCESS_CODE,
             self::SETTING_ENCAPSULATE_REVIEWERS,
             self::SETTING_EDITORS_CAN_REASSIGN_ARTICLES,
+            self::SETTING_AUTHORS_CAN_CONTACT_EDITORS,
+            self::SETTING_DISCLOSE_EDITOR_NAMES_TO_AUTHORS,
             self::SETTING_SYSTEM_AUTO_EDITORS_ASSIGNMENT,
             self::SETTING_AUTOMATICALLY_REASSIGN_SAME_REVIEWERS_WHEN_NEW_VERSION,
             self::SETTING_SYSTEM_NOTIFICATIONS,
@@ -1079,6 +1083,8 @@ class Episciences_Review
             self::SETTING_EDITORS_CAN_EDIT_TEMPLATES,
             self::SETTING_SYSTEM_AUTO_EDITORS_ASSIGNMENT,
             self::SETTING_EDITORS_CAN_ABANDON_CONTINUE_PUBLICATION_PROCESS,
+            self::SETTING_AUTHORS_CAN_CONTACT_EDITORS,
+            self::SETTING_DISCLOSE_EDITOR_NAMES_TO_AUTHORS,
         ], 'editors', ["legend" => "Paramètres des rédacteurs"]);
         $form->getDisplayGroup('editors')->removeDecorator('DtDdWrapper');
 
@@ -1462,6 +1468,19 @@ class Episciences_Review
                 'decorators' => $checkboxDecorators]
         );
 
+        $form->addElement('checkbox', self::SETTING_AUTHORS_CAN_CONTACT_EDITORS, [
+                'label' => "Permettre aux auteurs de contacter les rédacteurs responsables",
+                'description' => "Si activé, les auteurs peuvent envoyer des messages directement aux rédacteurs assignés à leur article",
+                'options' => ['uncheckedValue' => 0, 'checkedValue' => 1],
+                'decorators' => $checkboxDecorators]
+        );
+
+        $form->addElement('checkbox', self::SETTING_DISCLOSE_EDITOR_NAMES_TO_AUTHORS, [
+                'label' => "Afficher les noms des rédacteurs aux auteurs",
+                'description' => "Si activé, les auteurs peuvent voir les noms des rédacteurs assignés à leur article",
+                'options' => ['uncheckedValue' => 0, 'checkedValue' => 1],
+                'decorators' => $checkboxDecorators]
+        );
 
         $form->addElement('checkbox', self::SETTING_DO_NOT_ALLOW_EDITOR_IN_CHIEF_SELECTION, [
                 'label' => "Ne pas permettre le choix d'un rédacteur en chef",
@@ -1833,7 +1852,8 @@ class Episciences_Review
             self::SETTING_ARXIV_PAPER_PASSWORD, self::SETTING_DISPLAY_STATISTICS, self::SETTING_CONTACT_ERROR_MAIL,
             self::SETTING_REFUSED_ARTICLE_AUTHORS_MESSAGE_AUTOMATICALLY_SENT_TO_REVIEWERS,
             self::SETTING_TO_REQUIRE_REVISION_DEADLINE, self::SETTING_START_STATS_AFTER_DATE,
-            self::SETTING_ALLOW_EDIT_VOLUME_TITLE_WITH_PUBLISHED_ARTICLES, self::SETTING_DISPLAY_EMPTY_VOLUMES
+            self::SETTING_ALLOW_EDIT_VOLUME_TITLE_WITH_PUBLISHED_ARTICLES, self::SETTING_DISPLAY_EMPTY_VOLUMES,
+            self::SETTING_AUTHORS_CAN_CONTACT_EDITORS, self::SETTING_DISCLOSE_EDITOR_NAMES_TO_AUTHORS
         ];
 
         foreach ($settings as $setting) {
@@ -1856,7 +1876,42 @@ class Episciences_Review
         $allSettings[self::SETTING_JOURNAL_PUBLISHER] = trim(strip_tags((string)$this->getSetting(self::SETTING_JOURNAL_PUBLISHER)));
         $allSettings[self::SETTING_JOURNAL_PUBLISHER_LOC] = trim(strip_tags((string)$this->getSetting(self::SETTING_JOURNAL_PUBLISHER_LOC)));
 
-        if ($allSettings[self::SETTING_JOURNAL_PUBLISHER] === '' && $allSettings[self::SETTING_JOURNAL_PUBLISHER_LOC] !== '') {
+        // Permettre aux rédacteurs de demander des modifications sur un article
+        $settingsValues[self::SETTING_EDITORS_CAN_ASK_PAPER_REVISIONS] = $this->getSetting(self::SETTING_EDITORS_CAN_ASK_PAPER_REVISIONS);
+
+        // Permettre aux rédacteurs de modifier les templates
+        $settingsValues[self::SETTING_EDITORS_CAN_EDIT_TEMPLATES] = $this->getSetting(self::SETTING_EDITORS_CAN_EDIT_TEMPLATES);
+
+        // Permettre aux rédacteurs d'assigner des relecteurs à un article
+        //$settingsValues[self::SETTING_EDITORS_CAN_ASSIGN_REVIEWERS] = $this->getSetting(self::SETTING_EDITORS_CAN_ASSIGN_REVIEWERS);
+
+        // Permettre aux rédacteurs de choisir les rédacteurs d'un article
+        //$settingsValues[self::SETTING_EDITORS_CAN_ASSIGN_EDITORS] = $this->getSetting(self::SETTING_EDITORS_CAN_ASSIGN_EDITORS);
+
+        // Protection des volumes spéciaux par un code d'accès
+        $settingsValues[self::SETTING_SPECIAL_ISSUE_ACCESS_CODE] = $this->getSetting(self::SETTING_SPECIAL_ISSUE_ACCESS_CODE);
+
+        // Encapsulation des relecteurs au sein des volumes
+        $settingsValues[self::SETTING_ENCAPSULATE_REVIEWERS] = $this->getSetting(self::SETTING_ENCAPSULATE_REVIEWERS);
+
+        // Les rédacteurs peuvent réattribuer la gestion d'un article
+        $settingsValues[self::SETTING_EDITORS_CAN_REASSIGN_ARTICLES] = $this->getSetting(self::SETTING_EDITORS_CAN_REASSIGN_ARTICLES);
+
+        // Permettre aux auteurs de contacter les rédacteurs responsables
+        $settingsValues[self::SETTING_AUTHORS_CAN_CONTACT_EDITORS] = $this->getSetting(self::SETTING_AUTHORS_CAN_CONTACT_EDITORS);
+
+        // Divulguer les noms des rédacteurs aux auteurs
+        $settingsValues[self::SETTING_DISCLOSE_EDITOR_NAMES_TO_AUTHORS] = $this->getSetting(self::SETTING_DISCLOSE_EDITOR_NAMES_TO_AUTHORS);
+
+        // Encapsulation des préparateurs de copie
+        $settingsValues[self::SETTING_ENCAPSULATE_COPY_EDITORS] = $this->getSetting(self::SETTING_ENCAPSULATE_COPY_EDITORS);
+
+        // publisher
+        $settingsValues[self::SETTING_JOURNAL_PUBLISHER] = trim(strip_tags($this->getSetting(self::SETTING_JOURNAL_PUBLISHER)));
+
+        $settingsValues[self::SETTING_JOURNAL_PUBLISHER_LOC] = trim(strip_tags($this->getSetting(self::SETTING_JOURNAL_PUBLISHER_LOC)));
+
+        if ($settingsValues[self::SETTING_JOURNAL_PUBLISHER] === '' && $settingsValues[self::SETTING_JOURNAL_PUBLISHER_LOC] !== '') {
             return false;
         }